X86: Significantly filled out misc regs.
--HG--
extra : convert_revision : 4c53be6568134d65e57f5411df986fd9a89e82c9

+++ src/arch/x86/types.hh
@@ -257,7 +257,7 @@ namespace X86ISA
 
     typedef int RegContextVal;
 
-    typedef uint8_t RegIndex;
+    typedef uint16_t RegIndex;
 };
 
 #endif // __ARCH_X86_TYPES_HH__
